requirements.txt
numpy==1.18.2
Pillow==7.0.0
pkg-resources==0.0.0
